# docs/engineering/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Paperx",
    "description": "Stage 0 read-only feasibility API",
    "version": "0.1.0"
  },
  "paths": {
    "/health": {
      "get": {
        "summary": "Health",
        "operationId": "health_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/samples": {
      "get": {
        "summary": "Samples",
        "operationId": "samples_samples_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Sample"
                  },
                  "type": "array",
                  "title": "Response Samples Samples Get"
                }
              }
            }
          }
        }
      }
    },
    "/samples/{sample_id}/document": {
      "get": {
        "summary": "Document",
        "operationId": "document_samples__sample_id__document_get",
        "parameters": [
          {
            "name": "sample_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Sample Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Document"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        }
      }
    },
    "/samples/{sample_id}/pdf": {
      "get": {
        "summary": "Pdf",
        "operationId": "pdf_samples__sample_id__pdf_get",
        "parameters": [
          {
            "name": "sample_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Sample Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ApiError": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ApiError"
      },
      "BBox": {
        "properties": {
          "x": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "X"
          },
          "y": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Y"
          },
          "width": {
            "type": "number",
            "maximum": 1.0,
            "exclusiveMinimum": 0.0,
            "title": "Width"
          },
          "height": {
            "type": "number",
            "maximum": 1.0,
            "exclusiveMinimum": 0.0,
            "title": "Height"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "x",
          "y",
          "width",
          "height"
        ],
        "title": "BBox"
      },
      "Block": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "type": {
            "type": "string",
            "enum": [
              "heading",
              "paragraph",
              "formula",
              "figure",
              "table",
              "citation",
              "list",
              "code"
            ],
            "title": "Type"
          },
          "page": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Page"
          },
          "order": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Order"
          },
          "bbox": {
            "$ref": "#/components/schemas/BBox"
          },
          "locations": {
            "items": {
              "$ref": "#/components/schemas/Location"
            },
            "type": "array",
            "title": "Locations"
          },
          "text": {
            "type": "string",
            "title": "Text",
            "default": ""
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "confidence": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Confidence"
          },
          "heading_level": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 6.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Heading Level"
          },
          "formula": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Formula"
              },
              {
                "type": "null"
              }
            ]
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "type",
          "page",
          "order",
          "bbox",
          "source",
          "confidence"
        ],
        "title": "Block"
      },
      "Document": {
        "properties": {
          "schema_version": {
            "type": "string",
            "const": "1.0",
            "title": "Schema Version",
            "default": "1.0"
          },
          "paper_id": {
            "type": "string",
            "title": "Paper Id"
          },
          "source_sha256": {
            "type": "string",
            "title": "Source Sha256"
          },
          "parser_version": {
            "type": "string",
            "title": "Parser Version"
          },
          "coordinate_system": {
            "type": "string",
            "const": "normalized-unrotated-mediabox-top-left",
            "title": "Coordinate System",
            "default": "normalized-unrotated-mediabox-top-left"
          },
          "pages": {
            "items": {
              "$ref": "#/components/schemas/Page"
            },
            "type": "array",
            "title": "Pages"
          },
          "blocks": {
            "items": {
              "$ref": "#/components/schemas/Block"
            },
            "type": "array",
            "title": "Blocks"
          },
          "sections": {
            "items": {
              "$ref": "#/components/schemas/Section"
            },
            "type": "array",
            "title": "Sections"
          },
          "translations": {
            "items": {
              "$ref": "#/components/schemas/TranslationBlock"
            },
            "type": "array",
            "title": "Translations"
          },
          "translation_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Translation Version"
          },
          "previous_parser_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Previous Parser Version"
          },
          "previous_id_map": {
            "additionalProperties": {
              "items": {
                "type": "string"
              },
              "type": "array"
            },
            "type": "object",
            "title": "Previous Id Map"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "paper_id",
          "source_sha256",
          "parser_version",
          "pages",
          "blocks"
        ],
        "title": "Document"
      },
      "ErrorDetail": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "retryable": {
            "type": "boolean",
            "title": "Retryable",
            "default": false
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorDetail"
      },
      "Formula": {
        "properties": {
          "latex": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latex"
          },
          "number": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Number"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "confidence": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Confidence"
          },
          "status": {
            "type": "string",
            "enum": [
              "unrecognized",
              "needs_review",
              "verified"
            ],
            "title": "Status",
            "default": "unrecognized"
          },
          "revision": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Revision",
            "default": 0
          },
          "verified_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Verified By"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "source",
          "confidence"
        ],
        "title": "Formula"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "Location": {
        "properties": {
          "page": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Page"
          },
          "bbox": {
            "$ref": "#/components/schemas/BBox"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "page",
          "bbox"
        ],
        "title": "Location"
      },
      "Page": {
        "properties": {
          "number": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Number"
          },
          "width": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Width"
          },
          "height": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Height"
          },
          "rotation": {
            "type": "integer",
            "enum": [
              0,
              90,
              180,
              270
            ],
            "title": "Rotation",
            "default": 0
          },
          "text_status": {
            "type": "string",
            "enum": [
              "available",
              "missing"
            ],
            "title": "Text Status"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "number",
          "width",
          "height",
          "text_status"
        ],
        "title": "Page"
      },
      "Paper": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "source": {
            "type": "string",
            "enum": [
              "upload",
              "arxiv",
              "local",
              "fixture"
            ],
            "title": "Source"
          },
          "source_sha256": {
            "type": "string",
            "title": "Source Sha256"
          },
          "page_count": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Page Count"
          },
          "parser_version": {
            "type": "string",
            "title": "Parser Version"
          },
          "original_pdf_url": {
            "type": "string",
            "title": "Original Pdf Url"
          },
          "translated_pdf_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Translated Pdf Url"
          },
          "task_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Task Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "title",
          "source",
          "source_sha256",
          "page_count",
          "parser_version",
          "original_pdf_url"
        ],
        "title": "Paper"
      },
      "Sample": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "paper": {
            "$ref": "#/components/schemas/Paper"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "synthetic": {
            "type": "boolean",
            "title": "Synthetic"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "paper",
          "label",
          "synthetic"
        ],
        "title": "Sample"
      },
      "Section": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "block_id": {
            "type": "string",
            "title": "Block Id"
          },
          "level": {
            "type": "integer",
            "maximum": 6.0,
            "minimum": 1.0,
            "title": "Level"
          },
          "children": {
            "items": {
              "$ref": "#/components/schemas/Section"
            },
            "type": "array",
            "title": "Children"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "title",
          "block_id",
          "level"
        ],
        "title": "Section"
      },
      "TranslationBlock": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "source_block_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "minItems": 1,
            "title": "Source Block Ids"
          },
          "text": {
            "type": "string",
            "title": "Text"
          },
          "status": {
            "type": "string",
            "enum": [
              "translated",
              "preserved",
              "failed",
              "unmapped"
            ],
            "title": "Status"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "source_block_ids",
          "text",
          "status"
        ],
        "title": "TranslationBlock"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}